Safari-611.1.10/LayoutTests/http/tests/security/window-named-proto-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'document.body.innerHTML')
